Add a conda environment file
conda env create -f rise-example.conda-installation.yml

If you try to install the parts of the environment separately, 
the environment will not be solved, giving the "missing python 3.1" error.
This error is unstable - sometimes it is like that, sometimes it installs fine. 

With the environment file, there is no uncertainty - it will install every time.